Move signals out of site_config.py
This commit moves the signals into a separate signals.py file, and
moves the senders to apps.py.

NOTE
====

The post_migrate signal fires on the job_board app, and by default this
app's migrations run before the sites app's migrations run. This results
in gen_site_config_post_migrate() failing since the example.com site
does not yet exist. To circumvent, we move the job_board app below the
sites app in settings.py so that example.com gets created before we try
to query for it. We will need to revisit this in future to see what the
more appropriate thing to do is.
